setting up config handling (for find coordinates)
