+++ .gitignore
@@ -134,3 +134,4 @@ venv.bak/
 .mypy_cache/
 .dmypy.json
 dmypy.json
+papylio/gui/config.yml
